analyse top layer when using one click
