+++ app/javascript/components/map-v2/components/popup/selectors.js
@@ -2,50 +2,14 @@ import { createSelector, createStructuredSelector } from 'reselect';
 import isEmpty from 'lodash/isEmpty';
 import { reverseLatLng } from 'utils/geoms';
 
-import { getActiveDatasetsState, getActiveLayers } from '../../selectors';
+import {
+  getActiveDatasetsState,
+  filterInteractions,
+  getSelectedInteraction
+} from 'components/map-v2/selectors';
 
-const getSelected = state => state.popup.selected;
 const getSearch = state => state.location && state.location.search;
-const getInteractions = state => state.popup.interactions;
 const getLatLng = state => state.popup.latlng;
-
-export const filterInteractions = createSelector(
-  [getInteractions],
-  interactions => {
-    if (isEmpty(interactions)) return null;
-    return Object.values(interactions)
-      .filter(i => !isEmpty(i.data))
-      .map(i => ({
-        ...i
-      }));
-  }
-);
-
-export const getSelectedInteraction = createSelector(
-  [filterInteractions, getSelected, getActiveLayers],
-  (options, selected, layers) => {
-    if (isEmpty(options)) return null;
-    const layersWithoutBoundaries = layers.filter(
-      l => !l.isBoundary && !isEmpty(l.interactionConfig)
-    );
-    // if there is an article (icon layer) then choose that
-    let selectedData = options.find(o => o.article);
-    // if there is nothing selected get the top layer
-    if (!selected && !!layersWithoutBoundaries.length) {
-      selectedData = options.find(
-        o => o.value === layersWithoutBoundaries[0].id
-      );
-    }
-    // if only one layer then get that
-    if (!selectedData && options.length === 1) selectedData = options[0];
-    // otherwise get based on selected
-    if (!selectedData) selectedData = options.find(o => o.value === selected);
-    const layer =
-      selectedData && layers && layers.find(l => l.id === selectedData.id);
-
-    return { ...selectedData, layer };
-  }
-);
 
 export const getIsBoundary = createSelector(
   getSelectedInteraction,
