+++ app/javascript/components/map-v2/index.js
@@ -2,6 +2,7 @@ import { createElement, PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import moment from 'moment';
+import isEqual from 'lodash/isEqual';
 import { format } from 'd3-format';
 import startCase from 'lodash/startCase';
 
@@ -39,7 +40,10 @@ class MapContainer extends PureComponent {
       bbox,
       geostoreBbox,
       setMapSettings,
-      layerBbox
+      layerBbox,
+      selectedInteraction,
+      setAnalysisView,
+      oneClickAnalysisActive
     } = this.props;
 
     // update landsat basemap when changing zoom
@@ -63,6 +67,13 @@ class MapContainer extends PureComponent {
     // if geostore changes
     if (geostoreBbox && geostoreBbox !== prevProps.geostoreBbox) {
       setMapSettings({ bbox: geostoreBbox });
+    }
+    // set analysis view if interaction changes
+    if (
+      oneClickAnalysisActive &&
+      !isEqual(selectedInteraction, prevProps.selectedInteraction)
+    ) {
+      setAnalysisView(selectedInteraction);
     }
   }
 
@@ -98,17 +109,9 @@ class MapContainer extends PureComponent {
   };
 
   handleMapInteraction = ({ e, article, output, layer }) => {
-    const {
-      setAnalysisView,
-      setInteraction,
-      oneClickAnalysisActive,
-      draw
-    } = this.props;
-    const { data } = e || {};
+    const { setInteraction, draw } = this.props;
 
-    if (data && layer && oneClickAnalysisActive) {
-      setAnalysisView({ data, layer });
-    } else if (!draw) {
+    if (!draw) {
       setInteraction({
         ...e,
         label: layer.name,
@@ -143,7 +146,8 @@ MapContainer.propTypes = {
   oneClickAnalysisActive: PropTypes.bool,
   draw: PropTypes.bool,
   setAnalysisView: PropTypes.func,
-  setInteraction: PropTypes.func
+  setInteraction: PropTypes.func,
+  selectedInteraction: PropTypes.object
 };
 
 export default connect(getMapProps, actions)(MapContainer);
